+++ actionview/lib/action_view/railtie.rb
@@ -37,12 +37,6 @@ module ActionView
       end
     end
 
-    initializer "action_view.collection_caching" do |app|
-      ActiveSupport.on_load(:action_controller) do
-        PartialRenderer.collection_cache = app.config.action_controller.cache_store
-      end
-    end
-
     initializer "action_view.per_request_digest_cache" do |app|
       ActiveSupport.on_load(:action_view) do
         if app.config.consider_all_requests_local
@@ -55,6 +49,10 @@ module ActionView
       ActiveSupport.on_load(:action_controller) do
         ActionView::RoutingUrlFor.include(ActionDispatch::Routing::UrlFor)
       end
+    end
+
+    initializer "action_view.collection_caching", after: "action_controller.set_configs" do |app|
+      PartialRenderer.collection_cache = app.config.action_controller.cache_store
     end
 
     rake_tasks do |app|