+++ actionpack/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 5.0.0.beta4 (April 27, 2016) ##
+
 *   Routing: Refactor `:action` default handling to ensure that path
     parameters are not mutated during route generation.