+++ activemodel/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 5.0.0.beta4 (April 27, 2016) ##
+
 *   Allow passing record being validated to the message proc to generate
     customized error messages for that object using I18n helper.